fix: fix document.currentScript
