+++ dev/vue/src/main.js
@@ -134,3 +134,5 @@ setTimeout(() => {
 setTimeout(() => {
   throw new Error('vue app error');
 }, 3000);
+
+console.log(document.currentScript);
